Added more documentation and refactored out EitherPropertyMixin
